Add files via upload
The shopping data of people from Istanbul was retrieved from Kaggle, showing the shopping behaviour of males and females from 10 different malls. Using the mapper and reducer function of the Hadoop system we tried to understand the sales of different product categories across genders.  Also, we find the top sales for different product categories.